+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### Updated
 
+- Updated the Work/Alternate Address page to appear if the device location is in NYS but the Home Address is not in NYC.
 - Updated Design System to v0.18.6.
 - Updated copy to clarify the Work Address page and form fields.